requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
httpx==0.25.2
PyPDF2==3.0.1
python-docx==1.1.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.3.7
ollama==0.1.7